Enhance README with new Bedstack version announcement
This update adds a tip in the README to inform users about the release of a new version of _Bedstack (Stripped)_ that utilizes Drizzle v1. It includes a link to the `drizzle-v1` branch and notes that Drizzle v1 is currently in beta, with plans to merge into the main branch upon official release.